+++ src/main/java/com/basereh/app/StatisticMeasure.java
@@ -3,5 +3,5 @@ package com.basereh.app;
 import java.util.List;
 
 public interface StatisticMeasure {
-    public float calculate(List<Integer> nums);
+    public float calculate(List<Float> nums);
 }
